\name{ebSNP-package}
\alias{ebSNP-package}
\docType{package}
\title{
  Genotyping and SNP calling using single-sample next generation sequencing
data
}
\description{
  Genotyping and SNP calling for single-sample next generation
  sequencing using an empirical Bayes method
}
\details{
\tabular{ll}{
Package: \tab ebSNP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-02\cr
License: \tab \cr
}
}
\author{
Na You

Maintainer: Na You <youn@mail.sysu.edu.cn>
}
\references{
Single-sample SNP Detection By Empirical Bayes Method Using Next Generation Sequencing Data
}
\keyword{ package }
